feat: achieve complete MyPy type safety with 0 errors
This commit represents the culmination of systematic type safety improvements
across the entire codebase, eliminating all MyPy errors through proper
architectural solutions rather than suppressions.

Major improvements:
• TypedDict classes for structured data (GapFillerStats, GPUMemorySummary)
• Proper Union type handling with separate variables
• NumPy/pandas type compatibility fixes with explicit float() conversions
• Enhanced timezone-aware datetime operations
• Comprehensive type stubs for external libraries
• Strategic type casting for scikit-learn ambiguous returns

Key technical solutions:
• Fixed pandas DatetimeIndex comparisons with timezone-aware Timestamps
• Resolved NumPy floating type incompatibility through Python float conversion
• Implemented proper type narrowing for Union types
• Added comprehensive type annotations and forward references
• Eliminated dynamic attribute assignment conflicts

Files enhanced with proper type safety:
• ktrdr/ib/gap_filler.py - TypedDict for statistics structure
• ktrdr/training/gpu_memory_manager.py - Memory summary typing
• ktrdr/config/strategy_loader.py - Union type variable separation
• ktrdr/training/analytics/metrics_collector.py - scikit-learn type casting
• ktrdr/backtesting/performance.py - NumPy float compatibility
• ktrdr/api/services/gap_analysis_service.py - pandas timezone handling
• ktrdr/data/timeframe_synchronizer.py - Series operator workarounds
• ktrdr/training/train_strategy.py - metrics dictionary type safety

Documentation:
• Added comprehensive type-safety architecture document
• Outlined migration strategy for eliminating tool conflicts
• Provided blueprint for configuration-first design patterns

Result: 0 MyPy errors, proper type safety throughout codebase

+++ ktrdr/backtesting/performance.py
@@ -259,10 +259,14 @@ class PerformanceTracker:
                 profit_factor = 0.0
 
             avg_win = (
-                np.mean([t.net_pnl for t in winning_trades]) if winning_trades else 0
+                float(np.mean([t.net_pnl for t in winning_trades]))
+                if winning_trades
+                else 0.0
             )
             avg_loss = (
-                np.mean([t.net_pnl for t in losing_trades]) if losing_trades else 0
+                float(np.mean([t.net_pnl for t in losing_trades]))
+                if losing_trades
+                else 0.0
             )
             largest_win = (
                 max([t.net_pnl for t in winning_trades]) if winning_trades else 0
@@ -272,16 +276,18 @@ class PerformanceTracker:
             )
 
             # Holding period analysis
-            avg_holding_period = np.mean([t.holding_period_hours for t in trades])
+            avg_holding_period = float(
+                np.mean([t.holding_period_hours for t in trades])
+            )
             avg_win_holding_period = (
-                np.mean([t.holding_period_hours for t in winning_trades])
+                float(np.mean([t.holding_period_hours for t in winning_trades]))
                 if winning_trades
-                else 0
+                else 0.0
             )
             avg_loss_holding_period = (
-                np.mean([t.holding_period_hours for t in losing_trades])
+                float(np.mean([t.holding_period_hours for t in losing_trades]))
                 if losing_trades
-                else 0
+                else 0.0
             )
         else:
             total_trades = win_count = loss_count = 0
